add transaction to QuerySet.delete()
The delete method is unique in its order of operations
- create audit event(s) for delete operation
- call super().delete()
- save previously created audit events to DB

This means if make_audit_event raises an exception, there is no risk of gaps in the
audit log since no DB operations have been invoked yet. I still added a test where
make_audit_event was mocked to ensure this case is covered, but the more important
test is mocking the audit event save call, as if that fails, the delete operation
needs to be rolled back.