test saved attributes are in delta for delete event
This test used to assert using `instance.id`, but at this point
in the test, `instance` was deleted and therefore `instance.id`
was equivalent to None. Given that the instance was saved and the
id was set properly, this is not an accurate test.

Its past success was due to fetching the value to be used
in the delta for the delete audit event from the instance post deletion.
See https://github.com/dimagi/django-field-audit/blob/2206aeea2b71f51e1422ff5656e928a577cd8d8a/field_audit/models.py#L342

Its current succes is due to not setting init values
properly when creating a new instance of an audited model, which
meant the init values used when creating an audit event for the
deletion of the instance weren't correct.