Switch from nose to pynose
pynose is an updated version of nose, with support for newer
versions of python. The lack of support for newer versions of
python was initially a reason to switch test frameworks, but
pynose enables retaining the most compatibility with the current
test framework while also enabling usage of newer python versions.