Explicitly use models.Q(..., _negated=True) to prevent new migration
In Django 4.2, the behavior of __invert__ on Q objects was optimized
to toggle the negated attribute on the existing Q rather than add
an AND operator that wraps the inner clause.

On Django 4.2:
> from django.db import models
> ~(models.Q(test=1) | models.Q(test=2))
<Q: (NOT (OR: ('test', 1), ('test', 2)))>
> models.Q(models.Q(test=1) | models.Q(test=2), _negated=True)
<Q: (NOT (AND: (OR: ('test', 1), ('test', 2))))>

The optimized version is functionally the same, but creates an
unnecessary migration. This commit changes the syntax used when
defining the constraint to preserve the current format defined
in the migration.